requirements.txt
black==19.10b0
cffi==1.14.2
JACK-Client==0.5.2
psutil==5.7.2
pytest==6.0.1
watchgod==0.6
